Create centralized share menu component
