+++ frontend/src/features/event/results/attendees/mobile-drawer.tsx
@@ -4,14 +4,14 @@ import { ShareIcon, SquarePenIcon } from "lucide-react";
 
 import EmptyButton from "@/features/button/components/empty";
 import LinkButton from "@/features/button/components/link";
-import { FloatingDrawer, MorphingDrawer } from "@/features/drawer";
+import { MorphingDrawer } from "@/features/drawer";
 import PanelHeader from "@/features/event/results/attendees/panel-header";
 import ParticipantList from "@/features/event/results/attendees/participant-list";
 import {
   RemoveParticipantDialog,
   useParticipantRemoval,
 } from "@/features/event/results/attendees/remove-participant";
-import ShareMenu from "@/features/event/results/share-menu";
+import ShareMenu from "@/features/share-menu/menu";
 
 export default function AttendeesDrawer({
   onSnapChange,
@@ -61,9 +61,6 @@ export default function AttendeesDrawer({
     />
   );
 
-  /* SHARE MENU */
-  const [shareMenuOpen, setShareMenuOpen] = useState(false);
-
   return (
     <MorphingDrawer
       open
@@ -88,11 +85,7 @@ export default function AttendeesDrawer({
       }
       footerContent={
         <div className="mx-1 flex grow justify-between gap-2">
-          <FloatingDrawer
-            title="Share Event"
-            description="Share this event with others"
-            open={shareMenuOpen}
-            onOpenChange={setShareMenuOpen}
+          <ShareMenu
             trigger={
               <EmptyButton
                 buttonStyle="semi-transparent"
@@ -100,10 +93,10 @@ export default function AttendeesDrawer({
                 aria-label="Share Event"
               />
             }
-            nested={true}
-          >
-            <ShareMenu eventTitle={eventTitle} eventCode={eventCode} />
-          </FloatingDrawer>
+            eventTitle={eventTitle}
+            eventCode={eventCode}
+            isNested
+          />
           {paintingButton}
         </div>
       }
